% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pip.R
\name{plot_pip}
\alias{plot_pip}
\title{Pairwise Invasibility Plot}
\usage{
plot_pip(path, beta, s)
}
\arguments{
\item{path}{Path to where data for PIP are stored.}

\item{beta}{Choice parameter value.}

\item{s}{Selection coefficient value.}
}
\description{
This function plots a PIP for a given combination of parameters beta (choice parameter) and s (selection coefficient).
}
\author{
Raphael Scherrer
}
