% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singularities_along.R
\name{plot_singularities_along}
\alias{plot_singularities_along}
\title{Plot evolutionary singularities}
\usage{
plot_singularities_along(path = ".", var = "s", subset_by = "beta",
  subset_value = 0)
}
\arguments{
\item{path}{Path to where data for singularities are stored.}

\item{var}{Parameter along which to plot singularities.}

\item{subset_by}{Parameter for subsetting.}

\item{subset_value}{Value of the subset parameter.}
}
\description{
This function plots the singularities of the adaptive dynamics along one parameter, either the selection coefficient (s) or the choice parameter (beta). Data are subset by the other parameter.
}
\author{
Raphael Scherrer
}
